Fixed the signup page
