+++ src/pages/Signup.jsx
@@ -1,134 +1,63 @@
-import { useContext, useMemo, useState } from "react";
+import { useContext, useState } from "react";
 import { Link, useNavigate } from "react-router-dom";
 
 import {
   Typography,
   TextField,
   InputAdornment,
-  FormControlLabel,
-  Checkbox,
   Button,
   CircularProgress,
-  FormHelperText,
-  MenuItem,
   Box,
+  FormHelperText,
   useMediaQuery,
   useTheme,
   Toolbar,
 } from "@mui/material";
 
 import VisibilityOffIcon from "@mui/icons-material/VisibilityOff";
+import AppleIcon from "@mui/icons-material/Apple";
+import EmailIcon from "@mui/icons-material/Email";
 
 import { UtilsContext } from "../contexts/UtilsContext";
-import { signup } from "../utils/AuthFn";
-import { countries } from "../utils/Countries";
+import { login } from "../utils/AuthFn";
+import { googleLogin } from "../utils/AuthFn";
 import Navbar from "../components/Navbar";
-import axios from "axios";
 
 const Signup = () => {
-  const [firstName, setFirstName] = useState();
-  const [lastName, setLastName] = useState();
-  const [phoneNumber, setPhoneNumber] = useState();
   const [email, setEmail] = useState();
   const [password, setPassword] = useState();
-  const [confirmPassword, setConfirmPassword] = useState();
-  const [acceptTnC, setAcceptTnC] = useState(false);
-  const [acceptMU, setAcceptMU] = useState(false);
-  const [phoneErr, setPhoneErr] = useState(null);
   const [togglePassword, setTogglePassword] = useState("password");
   const [visibilityColor, setVisibilityColor] = useState("817e7e");
-  const [defaultCountry, setDefaultCountry] = useState("");
-  const [country, setCountry] = useState("");
-  const [countryCode, setCountryCode] = useState("");
 
-  const { errorMsg, loading } = useContext(UtilsContext);
+  const { loading } = useContext(UtilsContext);
+  const { errorMsg } = useContext(UtilsContext);
   const { setErrorMsg } = useContext(UtilsContext);
   const { setLoading } = useContext(UtilsContext);
   const navigate = useNavigate();
 
   const theme = useTheme();
-  console.log(theme);
   const md = useMediaQuery(theme.breakpoints.up("md"));
 
-  console.log(loading);
-  console.log({ country });
-
-  const handleClick = async (
-    firstName,
-    lastName,
-    email,
-    phoneNumber,
-    password,
-    confirmPassword
-  ) => {
-    try {
-      setLoading(true);
-      if (password !== confirmPassword) {
-        setErrorMsg({ confirmPassword: "Passwords are not the same" });
+  const handleClick = {
+    login: async (email, password) => {
+      try {
+        setLoading(true);
+        await login(email, password);
+        setLoading(false);
+        navigate("/dashboard");
+      } catch (error) {
+        setLoading(false);
+        setErrorMsg({ passwordLogin: "Incorrect email or password" });
         setTimeout(() => {
-          setErrorMsg({ confirmPassword: null });
-        }, 5000);
-        throw Error("Passwords are not the same");
-      }
-
-      await signup(firstName, lastName, email, phoneNumber, password);
-      setLoading(false);
-      navigate("/confirm-email");
-    } catch (error) {
-      setLoading(false);
-      if (error.message.includes("email")) {
-        setErrorMsg({ email: "This email is already in use" });
-        setTimeout(() => {
-          setErrorMsg({ email: null });
+          setErrorMsg({ passwordLogin: null });
         }, 5000);
       }
-
-      if (error.message.includes("password")) {
-        setErrorMsg({
-          password:
-            "Password must be at least 8 characters with 1 letter and 1 number",
-        });
-        setTimeout(() => {
-          setErrorMsg({ password: null });
-        }, 5000);
-      }
-    }
-  };
-
-  const handleChange = {
-    acceptTnC: function (e) {
-      setAcceptTnC(e.target.checked);
     },
 
-    acceptMU: function (e) {
-      setAcceptMU(e.target.checked);
-    },
-
-    changeCountry: function (e) {
-      setCountry(e.target.value);
-    },
-
-    setPhoneErr: function (length) {
-      console.log(length.length);
-      if (length.length < 5 || length.length > 14) {
-        setPhoneErr("Please enter a valid phone number.");
-      } else {
-        setPhoneErr(null);
-      }
+    googleLogin: async () => {
+      await googleLogin();
     },
   };
-
-  useMemo(async () => {
-    try {
-      const { data } = await axios.get("https://ipapi.co/json/");
-      setDefaultCountry(data.country_name);
-      const name = countries.find((item) => item.name === defaultCountry);
-      setCountryCode(name.mobileCode);
-      setCountry(defaultCountry);
-    } catch (error) {
-      return "Request failed";
-    }
-  }, [defaultCountry]);
 
   return (
     <Box>
@@ -136,17 +65,13 @@ const Signup = () => {
       <Toolbar sx={{ mb: 1 }} />
       {md ? (
         <Box
-          className="signupParent"
-          sx={{
-            position: "relative",
-            height: "150vh",
-            pb: 2,
-          }}
+          className="loginParent"
+          sx={{ position: "relative", height: "100vh" }}
         >
           <Box
             sx={{
               position: "relative",
-              height: "600px",
+              height: "400px",
               width: "80%",
               margin: "0px auto",
               backgroundColor: "#817e7e",
@@ -175,316 +100,155 @@ const Signup = () => {
               mx: 2,
               px: 4,
               py: 4,
-              mb: 4,
             }}
           >
             <Typography
               className="headingStyle"
               variant="h4"
-              sx={{ fontWeight: "bold", pb: 4 }}
+              sx={{ fontWeight: "bold", pb: 6 }}
             >
-              Create Personal Account
+              Payonize Sign up
             </Typography>
 
-            <form
-              onSubmit={(e) => {
-                e.preventDefault();
-                handleClick(
-                  firstName,
-                  lastName,
-                  email,
-                  phoneNumber,
-                  password,
-                  confirmPassword
-                );
+            <Button
+              className="buttonStyle"
+              variant="contained"
+              fullWidth
+              sx={{
+                py: 1.5,
+                px: 3,
+                mt: 1,
+                mb: 2,
+                fontWeight: "bold",
+                textTransform: "capitalize",
+                background: "#b4b2b2",
+                color: "#000000",
+                ":hover": {
+                  background: "#b4b2b2",
+                },
+              }}
+              onClick={async () => {
+                await handleClick.googleLogin();
               }}
             >
-              <label>First Name</label>
-              <TextField
-                type="text"
-                fullWidth
-                required
-                size="small"
-                sx={{ mt: 1, mb: 2 }}
-                onChange={(e) => {
-                  setFirstName(e.target.value);
-                }}
-                inputProps={{
-                  minLength: 3,
-                }}
-              />
-
-              <label>Last Name</label>
-              <TextField
-                type="text"
-                fullWidth
-                required
-                size="small"
-                sx={{ mt: 1, mb: 2 }}
-                onChange={(e) => {
-                  setLastName(e.target.value);
-                }}
-                inputProps={{
-                  minLength: 3,
-                }}
-              />
-
-              <label>Email</label>
-              <TextField
-                type="email"
-                fullWidth
-                required
-                size="small"
-                sx={{ mt: 1 }}
-                onChange={(e) => {
-                  setEmail(e.target.value);
-                }}
-              />
-              <FormHelperText error sx={{ mb: 2 }}>
-                {errorMsg.email}
-              </FormHelperText>
-
-              <label>Phone Number</label>
-
-              <Box display={"flex"}>
-                <TextField
-                  select
-                  sx={{ mt: 1, mr: 2 }}
-                  value={country}
-                  onChange={handleChange.changeCountry}
-                  size="small"
-                >
-                  {countries.map((country) => (
-                    <MenuItem value={country.name} key={country.name}>
-                      {country.mobileCode}: {country.code}
-                    </MenuItem>
-                  ))}
-                </TextField>
-
-                <TextField
-                  type="number"
-                  fullWidth
-                  required
-                  size="small"
-                  sx={{ mt: 1 }}
-                  onChange={(e) => {
-                    const digits = e.target.valueAsNumber;
-                    setPhoneNumber(`${countryCode}${digits}`);
-                    handleChange.setPhoneErr(e.target.value);
-                    console.log(phoneNumber);
-                  }}
-                  onMouseOut={(e) => {}}
+              <Box display="flex" alignItems="center" justifyContent="center">
+                <img
+                  src="google.png"
+                  alt="Google"
+                  style={{ width: "25px", height: "25px" }}
                 />
+                <span style={{ marginLeft: "8px" }}>Continue with Google</span>
               </Box>
-              <FormHelperText error sx={{ mb: 2 }}>
-                {phoneErr}
-              </FormHelperText>
+            </Button>
 
-              <label>Password</label>
-              <TextField
-                type={togglePassword}
-                fullWidth
-                required
-                size="small"
-                sx={{ mt: 1 }}
-                InputProps={{
-                  endAdornment: (
-                    <InputAdornment
-                      position="end"
-                      sx={{ cursor: "pointer" }}
-                      onClick={() => {
-                        if (togglePassword === "password") {
-                          setTogglePassword("text");
-                          setVisibilityColor("54adf3");
-                        } else {
-                          setTogglePassword("password");
-                          setVisibilityColor("817e7e");
-                        }
-                      }}
-                    >
-                      <VisibilityOffIcon
-                        sx={{ color: `#${visibilityColor}` }}
-                      />
-                    </InputAdornment>
-                  ),
-                }}
-                onChange={(e) => setPassword(e.target.value)}
-              />
-              <FormHelperText error sx={{ mb: 2 }}>
-                {errorMsg.password}
-              </FormHelperText>
+            <Button
+              className="buttonStyle"
+              variant="contained"
+              fullWidth
+              sx={{
+                py: 1.5,
+                px: 3,
+                mt: 3,
+                mb: 2,
+                textTransform: "capitalize",
+                fontWeight: "bold",
+                background: "#b4b2b2",
+                color: "#000000",
+                ":hover": {
+                  background: "#b4b2b2",
+                },
+              }}
+            >
+              <Box display="flex" alignItems="center" justifyContent="center">
+                <AppleIcon sx={{ mr: 1 }} /> <span>Continue with Apple</span>
+              </Box>
+            </Button>
 
-              <label>Confirm Password</label>
-              <TextField
-                type="password"
-                fullWidth
-                required
-                size="small"
-                sx={{ mt: 1 }}
-                onChange={(e) => setConfirmPassword(e.target.value)}
-              />
-              <FormHelperText error sx={{ mb: 2 }}>
-                {errorMsg.confirmPassword}
-              </FormHelperText>
-
-              <FormControlLabel
-                label="I have read and agreed to Payonize Terms of Service and Privacy Policy"
-                control={
-                  <Checkbox
-                    checked={acceptTnC}
-                    onChange={handleChange.acceptTnC}
-                  />
-                }
-              />
-
-              <FormControlLabel
-                label="I agree to receive marketing updates from Payonize"
-                control={
-                  <Checkbox
-                    checked={acceptMU}
-                    onChange={handleChange.acceptMU}
-                  />
-                }
-              />
-
-              <Button
-                className="buttonStyle"
-                type="submit"
-                variant="contained"
-                fullWidth
-                disabled={acceptTnC === false || loading || phoneErr !== null}
+            <Box
+              display="flex"
+              alignItems="center"
+              justifyContent="space-between"
+              sx={{
+                my: 3,
+              }}
+            >
+              <Box
                 sx={{
-                  py: 1.5,
-                  px: 3,
-                  mt: 2,
-                  fontWeight: "bold",
-                  textTransform: "capitalize",
-                  color: "#000000",
-                  background: "#54adf3",
-                  ":hover": {
-                    background: "#54adf3",
-                  },
+                  background: "#817e7e",
+                  height: "1px",
+                  flex: 1,
                 }}
-              >
-                {loading ? (
-                  <CircularProgress size={"1.5rem"} sx={{ color: "#ffffff" }} />
-                ) : (
-                  "Create Personal Account"
-                )}
-              </Button>
-            </form>
+              ></Box>
+              <Box sx={{ px: 2 }}>or</Box>
+              <Box
+                sx={{
+                  background: "#817e7e",
+                  height: "1px",
+                  flex: 1,
+                }}
+              ></Box>
+            </Box>
 
-            <Typography sx={{ mt: 2 }}>
+            <Button
+              className="buttonStyle"
+              type="submit"
+              variant="contained"
+              fullWidth
+              sx={{
+                py: 1.5,
+                px: 3,
+                mt: 2,
+                mb: 1,
+                fontWeight: "bold",
+                textTransform: "capitalize",
+                background: "#b4b2b2",
+                color: "#000000",
+                ":hover": {
+                  background: "#b4b2b2",
+                },
+              }}
+              onClick={() => navigate("/signup/email")}
+            >
+              <Box display="flex" alignItems="center" justifyContent="center">
+                <EmailIcon sx={{ mr: 1 }} /> <span>Continue with Email</span>
+              </Box>
+            </Button>
+
+            <Typography sx={{ mt: 2, mb: 4 }}>
               Already have an account?{" "}
               <Link to="/login" style={{ color: "#54adf3" }}>
-                Log In
+                Login
               </Link>
             </Typography>
           </Box>
         </Box>
       ) : (
-        <Box sx={{ py: 2, px: 4 }}>
+        <Box sx={{ py: 2, px: 2 }}>
           <Typography
             className="headingStyle"
             variant="h4"
             sx={{ fontWeight: "bold", pb: 4 }}
           >
-            Create Personal Account
+            Payonize Login
           </Typography>
 
           <form
             onSubmit={(e) => {
               e.preventDefault();
-              handleClick(
-                firstName,
-                lastName,
-                email,
-                phoneNumber,
-                password,
-                confirmPassword
-              );
+              handleClick.login(email, password);
             }}
           >
-            <label>First Name</label>
-            <TextField
-              type="text"
-              fullWidth
-              required
-              size="small"
-              sx={{ mt: 1, mb: 2 }}
-              onChange={(e) => {
-                setFirstName(e.target.value);
-              }}
-              inputProps={{
-                minLength: 3,
-              }}
-            />
-
-            <label>Last Name</label>
-            <TextField
-              type="text"
-              fullWidth
-              required
-              size="small"
-              sx={{ mt: 1, mb: 2 }}
-              onChange={(e) => {
-                setLastName(e.target.value);
-              }}
-              inputProps={{
-                minLength: 3,
-              }}
-            />
-
             <label>Email</label>
             <TextField
               type="email"
               fullWidth
               required
               size="small"
-              sx={{ mt: 1 }}
               onChange={(e) => {
                 setEmail(e.target.value);
               }}
+              sx={{ mb: 2 }}
             />
-            <FormHelperText error sx={{ mb: 2 }}>
-              {errorMsg.email}
-            </FormHelperText>
-
-            <label>Phone Number</label>
-
-            <Box display={"flex"}>
-              <TextField
-                select
-                sx={{ mt: 1, mr: 2 }}
-                value={country}
-                onChange={handleChange.changeCountry}
-                size="small"
-              >
-                {countries.map((country) => (
-                  <MenuItem value={country.name} key={country.name}>
-                    {country.mobileCode}: {country.code}
-                  </MenuItem>
-                ))}
-              </TextField>
-
-              <TextField
-                type="number"
-                fullWidth
-                required
-                size="small"
-                sx={{ mt: 1 }}
-                onChange={(e) => {
-                  const digits = e.target.valueAsNumber;
-                  setPhoneNumber(`${countryCode}${digits}`);
-                  handleChange.setPhoneErr(e.target.value);
-                  console.log(phoneNumber);
-                }}
-                onMouseOut={(e) => {}}
-              />
-            </Box>
-            <FormHelperText error sx={{ mb: 2 }}>
-              {phoneErr}
-            </FormHelperText>
 
             <label>Password</label>
             <TextField
@@ -515,49 +279,20 @@ const Signup = () => {
               onChange={(e) => setPassword(e.target.value)}
             />
             <FormHelperText error sx={{ mb: 2 }}>
-              {errorMsg.password}
+              {errorMsg.passwordLogin}
             </FormHelperText>
-
-            <label>Confirm Password</label>
-            <TextField
-              type="password"
-              fullWidth
-              required
-              size="small"
-              sx={{ mt: 1 }}
-              onChange={(e) => setConfirmPassword(e.target.value)}
-            />
-            <FormHelperText error sx={{ mb: 2 }}>
-              {errorMsg.confirmPassword}
-            </FormHelperText>
-
-            <FormControlLabel
-              label="I have read and agreed to Payonize Terms of Service and Privacy Policy"
-              control={
-                <Checkbox
-                  checked={acceptTnC}
-                  onChange={handleChange.acceptTnC}
-                />
-              }
-            />
-
-            <FormControlLabel
-              label="I agree to receive marketing updates from Payonize"
-              control={
-                <Checkbox checked={acceptMU} onChange={handleChange.acceptMU} />
-              }
-            />
 
             <Button
               className="buttonStyle"
               type="submit"
               variant="contained"
               fullWidth
-              disabled={acceptTnC === false || loading || phoneErr !== null}
+              disabled={loading}
               sx={{
                 py: 1.5,
                 px: 3,
-                mt: 2,
+                mt: 1,
+                mb: 1,
                 fontWeight: "bold",
                 textTransform: "capitalize",
                 color: "#000000",
@@ -570,15 +305,90 @@ const Signup = () => {
               {loading ? (
                 <CircularProgress size={"1.5rem"} sx={{ color: "#ffffff" }} />
               ) : (
-                "Create Personal Account"
+                "Login"
               )}
             </Button>
           </form>
 
+          <Box
+            display="flex"
+            alignItems="center"
+            justifyContent="space-between"
+            sx={{
+              my: 2,
+            }}
+          >
+            <Box
+              sx={{
+                background: "#817e7e",
+                height: "1px",
+                flex: 1,
+              }}
+            ></Box>
+            <Box sx={{ px: 2 }}>or</Box>
+            <Box
+              sx={{
+                background: "#817e7e",
+                height: "1px",
+                flex: 1,
+              }}
+            ></Box>
+          </Box>
+
+          <Button
+            className="buttonStyle"
+            variant="contained"
+            fullWidth
+            sx={{
+              py: 1.5,
+              px: 3,
+              mt: 1,
+              textTransform: "capitalize",
+              background: "#b4b2b2",
+              color: "#000000",
+              ":hover": {
+                background: "#b4b2b2",
+              },
+            }}
+            onClick={async () => {
+              await handleClick.googleLogin();
+            }}
+          >
+            <Box display="flex" alignItems="center" justifyContent="center">
+              <img
+                src="google.png"
+                alt="Nigeria"
+                style={{ width: "25px", height: "25px" }}
+              />
+              <span style={{ marginLeft: "8px" }}>Continue with Google</span>
+            </Box>
+          </Button>
+
+          <Button
+            className="buttonStyle"
+            variant="contained"
+            fullWidth
+            sx={{
+              py: 1.5,
+              px: 3,
+              mt: 3,
+              textTransform: "capitalize",
+              background: "#b4b2b2",
+              color: "#000000",
+              ":hover": {
+                background: "#b4b2b2",
+              },
+            }}
+          >
+            <Box display="flex" alignItems="center" justifyContent="center">
+              <AppleIcon sx={{ mr: 1 }} /> <span>Continue with Apple</span>
+            </Box>
+          </Button>
+
           <Typography sx={{ mt: 2 }}>
-            Already have an account?{" "}
-            <Link to="/login" style={{ color: "#54adf3" }}>
-              Log In
+            Don&apos;t have an account?{" "}
+            <Link to="/signup" style={{ color: "#54adf3" }}>
+              Sign up
             </Link>
           </Typography>
         </Box>
